Challenge 4:  Football Scripts
In this new scene, we have some problems to fix in order to be able to score goals while trying to avoid getting scored.

First of all, we improved the way we collided with the enemy because he was the one kicking us instead of us kicking him.

Second, the powerup was not working as intended and enemy spawning wasn´t showing the correct amount of objects.

Third, the enemies didn´t move at all towards the goal, so we put some movement and now with each wave, enemy speed is increased.

Fourth, there is now a boost element when pressing the spacebar key to move faster, which also includes a particle effect next to the focal Point element.